backend/requirements.txt
asgiref==3.8.1
attrs==25.4.0
bcrypt==5.0.0
beautifulsoup4==4.14.3
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.6
Django==5.2.1
django-cors-headers==4.9.0
django-filter==25.2
django-otp==1.7.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
et_xmlfile==2.0.0
greenlet==3.3.1
gunicorn==23.0.0
idna==3.11
inflection==0.5.1
invoke==3.0.0
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
numpy==2.3.4
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
paramiko==4.0.0
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.0.0
playwright==1.58.0
ply==3.11
psycopg2==2.9.11
puresnmp==2.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==3.0
pycryptodomex==3.23.0
pyee==13.0.1
PyJWT==2.10.1
PyNaCl==1.6.2
pyotp==2.9.0
pypdfium2==5.6.0
pysmi==1.6.2
pysnmp==7.1.22
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.2
pytz==2025.2
PyYAML==6.0.3
qrcode==8.2
referencing==0.37.0
reportlab==4.4.4
requests==2.32.5
rpds-py==0.30.0
scp==0.15.0
six==1.17.0
soupsieve==2.8.3
sqlparse==0.5.3
t61codec==1.0.1
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
x690==1.0.0.post1
dj-database-url==2.1.0
